+++ app/Werewolf/Commands/Start.hs
@@ -21,8 +21,8 @@ module Werewolf.Commands.Start (
 
 import Control.Monad.Except
 import Control.Monad.Extra
-import Control.Monad.Writer
 import Control.Monad.State
+import Control.Monad.Writer
 
 import Data.Text (Text)