+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Added missing apostrophe to the new turn message. ([#63](https://github.com/hjwylde/werewolf/issues/63))
 * Changed the "Whom would you like to lynch?" text to be a public message displayed after the devoured message. ([#56](https://github.com/hjwylde/werewolf/issues/56))
+* Better prompt to action when villagers vote.
 
 #### v0.3.3.2