+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added missing apostrophe to the new turn message. ([#63](https://github.com/hjwylde/werewolf/issues/63))
 * Changed the "Whom would you like to lynch?" text to be a public message displayed after the devoured message. ([#56](https://github.com/hjwylde/werewolf/issues/56))
 * Better prompt to action when villagers vote.
+* Changed devour vote messages to be sent immediately. ([#57](https://github.com/hjwylde/werewolf/issues/57))
 
 #### v0.3.3.2