+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Better prompt to action when villagers vote.
 * Changed devour vote messages to be sent immediately. ([#57](https://github.com/hjwylde/werewolf/issues/57))
 * Removed useless `only` function. ([#55](https://github.com/hjwylde/werewolf/issues/55))
+* Turned start of day and night into distinct turns.
 
 #### v0.3.3.2